OctreeFarm: Add Source::posToPointer() method

+++ Tools/OctreeFarm/Octree.cpp
@@ -112,7 +112,7 @@ int Octree::colorAttachAddress(int parent, int childIndex) {
     return attachData + (paletteNode >> 8) + Utils::bitCount8(paletteNode & ((1 << childIndex) - 1));
 }
 
-void Octree::select(uint32_t parent, uint32_t childIndex, const glm::vec3& pos, bool append) {
+void Octree::select(uint32_t parent, uint32_t childIndex, const glm::uvec3& pos, bool append) {
     int address = colorAttachAddress(parent, childIndex);
 
     QSharedPointer<Node> node;
